% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1measures.R
\name{step1measures}
\alias{step1measures}
\title{Compute Measures for Identifying Patterns of Change in Functional Data}
\usage{
step1measures(Data, Time = NULL, ID = FALSE, measures = 1:23, midpoint = NULL)
}
\arguments{
\item{Data}{A matrix or data frame in which each row contains the (ordered) observations of a given function.}

\item{Time}{Either NULL or a vector or a matrix (or data frame) of the same dimension as \code{Data}. If a vector, matrix or data frame is supplied, it is assumed that the entries correspond to the values of the arguments at which the functions in \code{Data} are observed. When set to \code{NULL} (the default), the observations are assumed equidistant.}

\item{ID}{Logical. Set to \code{TRUE} if the first column of \code{Data} and \code{Time} corresponds to an \code{ID} variable. Defaults to \code{FALSE}.}

\item{measures}{A numerical vector containing the numerical identifiers of the measures to compute (see "Details" section below). The default, c(1:23), leaves out the measures which require specifying a midpoint.}

\item{midpoint}{Specifies which column of \code{Time} to use as the midpoint in measures 24-26. Can be an integer, an integer vector (of length the number of rows in \code{Time}) or NULL (the default). If NULL, the midpoint for a given function is taken to be the time closest to the median.}
}
\value{
An object of class \code{trajMeasures}; a list containing the values of the measures, a table of outliers which have been imputed, as well as a curated form of the \code{Data} and \code{Time} arguments.
}
\description{
\code{step1measures} computes up to 26 measures for each function. See details for the list of measures.
}
\details{
Each function must have a minimum of 3 observations otherwise it will be omitted from the analysis.

The 26 measures (and their numerical identifiers) are:
\enumerate{
\item  Range\cr
\item  Mean of the function\cr
\item  Functional standard deviation (SD)\cr
\item  Coefficient of variation (ratio of measure 3 to measure 2)\cr
\item  Overall change (initial value - final value)\cr
\item  Mean change per unit time\cr
\item  Overall change relative to initial value\cr
\item  Overall change relative to functional mean (ratio of measure 5 to measure 5)\cr
\item  Slope of the linear model\cr
\item  \eqn{R^2}: Proportion of variance explained by the linear model\cr
\item  Maximum value of the speed\cr
\item  Functional SD of the speed\cr
\item  Mean absolute speed\cr
\item  Maximum absolute speed\cr
\item  Maximum absolute speed relative to the functional mean (ratio of measure 14 to measure 2)\cr
\item  Maximum absolute speed relative to the slope (ratio of measure 14 to measure 9)\cr
\item  Functional SD of the speed relative to the slope\cr(ratio of measure 12 to measure 9)
\item  Mean acceleration\cr
\item  Mean absolute acceleration\cr
\item  Maximum of the absolute acceleration\cr
\item  Maximum of the absolute acceleration relative to the functional (ratio of measure 20 to measure 2)\cr
\item  Maximum of the absolute acceleration relative to the mean absolute speed (ratio of measure 20 to measure 13)\cr
\item  Mean absolute acceleration relative to the mean absolute speed (ratio of measure 19 to measure 13)\cr
\item  Early change relative to later change\cr
\item  Early change relative to overall change\cr
\item  Later change relative to overall change\cr
}
}
\examples{
\dontrun{
data = example.data$data

m1 = step1measures(data, ID=TRUE, measures=24:26, midpoint=NULL)
m2 = step1measures(data, ID=TRUE, measures=24:26, midpoint=as.integer(3))

identical(s1$measures,s2$measures)
}

}
\references{
Leffondre K, Abrahamowicz M, Regeasse A, Hawker GA, Badley EM, McCusker J, Belzile E. Statistical measures were proposed for identifying longitudinal patterns of change in quantitative health indicators. J Clin Epidemiol. 2004 Oct;57(10):1049-62. doi: 10.1016/j.jclinepi.2004.02.012. PMID: 15528056.
}
