% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Step3Clusters.R
\name{Step3Clusters}
\alias{Step3Clusters}
\title{Classify the Longitudinal Data Based on the Selected Measures.}
\usage{
Step3Clusters(
  trajSelection,
  forced.measures = NULL,
  nstart = 50,
  iter.max = 20,
  nclusters = NULL,
  K.max = 8,
  B = 500,
  d.power = 2,
  spaceH0 = "scaledPCA",
  method = "Tibs2001SEmax",
  SE.factor = 1
)
}
\arguments{
\item{trajSelection}{Object of class \code{trajSelection} as returned by
\code{Step2Selection}.}

\item{forced.measures}{Either NULL or a numerical vectors corresponding to the
measures used for clustering. If NULL, the measures selected by
\code{Step2Selection} are used.}

\item{nstart}{to be passed to the nstart argument of
\code{\link[stats]{kmeans}}.}

\item{iter.max}{to be passed to the iter.max argument of
\code{\link[stats]{kmeans}}.}

\item{nclusters}{either NULL or the number of clusters to form. If NULL, the
number of clusters to form will be determined using the GAP criterion as
implemented in the \code{\link[cluster]{clusGap}} function.}

\item{K.max}{to be passed to the \code{K.max} argument of
\code{\link[cluster]{clusGap}}.}

\item{B}{to be passed to the \code{SE.factor} argument of
\code{\link[cluster]{clusGap}}.}

\item{d.power}{to be passed to the \code{B} argument of
\code{\link[cluster]{clusGap}}.}

\item{spaceH0}{to be passed to the \code{spaceH0} argument of
\code{\link[cluster]{clusGap}}.}

\item{method}{to be passed to the \code{method} argument of
\code{\link[cluster]{clusGap}}.}

\item{SE.factor}{to be passed to the \code{SE.factor} argument of
\code{\link[cluster]{clusGap}}.}
}
\value{
An object of class \code{trajClusters}; a list containing the result
of the clustering as well as curated forms of the data and time matrices.
}
\description{
Classifies the trajectories by applying the K-mean clustering
algorithm to the measures selected by \code{Step2Selection}.
}
\examples{
\dontrun{
data = example.data$data

m = Step1Measures(data, ID=TRUE)
s = Step2Selection(m)

s$loadings

s2 = Step2Selection(m,select=c(10,12,8,4))

groups3Cl <- step3(s2,nclusters=3)$partition
groups4Cl <- step3(s2,nclusters=4)$partition
groups5Cl <- step3(s2,nclusters=5)$partition

}

}
\seealso{
\code{\link[traj2]{Step2Selection}}
}
