% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traj2-package.R
\docType{package}
\name{traj2-package}
\alias{traj2}
\alias{traj2-package}
\title{traj2: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
